Fix Copilot review: enable RLS on upvotes table, sanitize search, capture errors, FLAG_MAP, centralize point values
