+++ lib/reputation/awards.ts
@@ -1,23 +1,10 @@
 import "server-only";
 import { createServiceClient } from "@/lib/supabase/server";
+import { REPUTATION_POINTS, NOTE_UPVOTE_CAP, type ReputationEventType } from "./points";
 
-// ─── Point values ────────────────────────────────────────────────────────────
-
-export const REPUTATION_POINTS = {
-  matter_joined: 30,
-  brief_generated: 50,
-  note_contributed: 40,
-  note_upvoted: 10,
-  handoff_completed: 25,
-  match_accepted: 20,
-  profile_completed: 60,
-  cross_border_matter: 100,
-} as const;
-
-export type ReputationEventType = keyof typeof REPUTATION_POINTS;
-
-// Max points a single field note can earn its author from upvotes
-export const NOTE_UPVOTE_CAP = 200;
+// Re-export for back-compat with existing server-side imports.
+export { REPUTATION_POINTS, NOTE_UPVOTE_CAP };
+export type { ReputationEventType };
 
 // ─── Badge tiers (re-exported from client-safe badges.ts) ───────────────────
 export { BADGE_TIERS, getBadge } from "@/lib/reputation/badges";
